requirements.txt
openai>=1.30.0
anthropic>=0.28.0
Pillow>=10.0.0
pytesseract>=0.3.10
editdistance>=0.8.1
scikit-learn>=1.4.0
tabulate>=0.9.0
python-dotenv>=1.0.0
numpy>=1.26.0
